+++ packages/components/src/internal/components/entities/EntityLineageEditModal.spec.tsx
@@ -71,7 +71,7 @@ const DEFAULT_PROPS = {
     parentEntityDataTypes: [SampleTypeDataType, DataClassDataType],
     api: getTestAPIWrapper(jest.fn, {
         samples: getSamplesTestAPIWrapper(jest.fn, {
-            getSampleSelectionLineageData: () => Promise.resolve(LINEAGE_DATA_WITHOUT_ALIQUOTS),
+            getSelectionLineageData: () => Promise.resolve(LINEAGE_DATA_WITHOUT_ALIQUOTS),
         }),
     }),
 };
@@ -150,7 +150,7 @@ describe('EntityLineageEditModal', () => {
                 queryModel={MODEL}
                 api={getTestAPIWrapper(jest.fn, {
                     samples: getSamplesTestAPIWrapper(jest.fn, {
-                        getSampleSelectionLineageData: () => Promise.resolve(LINEAGE_DATA_WITHOUT_ALIQUOTS),
+                        getSelectionLineageData: () => Promise.resolve(LINEAGE_DATA_WITHOUT_ALIQUOTS),
                         getSampleOperationConfirmationData: () =>
                             Promise.resolve(new OperationConfirmationData(ALL_ALLOWED_CONFIRMATION_DATA)),
                     }),
@@ -171,7 +171,7 @@ describe('EntityLineageEditModal', () => {
                 queryModel={MODEL}
                 api={getTestAPIWrapper(jest.fn, {
                     samples: getSamplesTestAPIWrapper(jest.fn, {
-                        getSampleSelectionLineageData: () => Promise.resolve(LINEAGE_DATA_WITH_ALIQUOTS),
+                        getSelectionLineageData: () => Promise.resolve(LINEAGE_DATA_WITH_ALIQUOTS),
                         getSampleOperationConfirmationData: () =>
                             Promise.resolve(new OperationConfirmationData(ALL_ALLOWED_CONFIRMATION_DATA)),
                     }),
@@ -195,7 +195,7 @@ describe('EntityLineageEditModal', () => {
                 queryModel={MODEL}
                 api={getTestAPIWrapper(jest.fn, {
                     samples: getSamplesTestAPIWrapper(jest.fn, {
-                        getSampleSelectionLineageData: () => Promise.resolve(LINEAGE_DATA_ALL_ALIQUOTS),
+                        getSelectionLineageData: () => Promise.resolve(LINEAGE_DATA_ALL_ALIQUOTS),
                         getSampleOperationConfirmationData: () =>
                             Promise.resolve(new OperationConfirmationData(ALL_ALLOWED_CONFIRMATION_DATA)),
                     }),
@@ -218,7 +218,7 @@ describe('EntityLineageEditModal', () => {
                 parentEntityDataTypes={[DataClassDataType, SampleTypeDataType]}
                 api={getTestAPIWrapper(jest.fn, {
                     samples: getSamplesTestAPIWrapper(jest.fn, {
-                        getSampleSelectionLineageData: () => Promise.resolve(LINEAGE_DATA_WITHOUT_ALIQUOTS),
+                        getSelectionLineageData: () => Promise.resolve(LINEAGE_DATA_WITHOUT_ALIQUOTS),
                         getSampleOperationConfirmationData: () =>
                             Promise.resolve(new OperationConfirmationData(ALL_ALLOWED_CONFIRMATION_DATA)),
                     }),
@@ -240,7 +240,7 @@ describe('EntityLineageEditModal', () => {
                 parentEntityDataTypes={[DataClassDataType, SampleTypeDataType]}
                 api={getTestAPIWrapper(jest.fn, {
                     samples: getSamplesTestAPIWrapper(jest.fn, {
-                        getSampleSelectionLineageData: () => Promise.resolve(LINEAGE_DATA_WITHOUT_ALIQUOTS),
+                        getSelectionLineageData: () => Promise.resolve(LINEAGE_DATA_WITHOUT_ALIQUOTS),
                         getSampleOperationConfirmationData: () =>
                             Promise.resolve(new OperationConfirmationData(NONE_ALLOWED_CONFIRMATION_DATA)),
                     }),
@@ -265,7 +265,7 @@ describe('EntityLineageEditModal', () => {
                 parentEntityDataTypes={[DataClassDataType, SampleTypeDataType]}
                 api={getTestAPIWrapper(jest.fn, {
                     samples: getSamplesTestAPIWrapper(jest.fn, {
-                        getSampleSelectionLineageData: () => Promise.resolve(LINEAGE_DATA_WITH_ALIQUOTS),
+                        getSelectionLineageData: () => Promise.resolve(LINEAGE_DATA_WITH_ALIQUOTS),
                         getSampleOperationConfirmationData: () =>
                             Promise.resolve(new OperationConfirmationData(SOME_ALLOWED_CONFIRMATION_DATA)),
                     }),
@@ -292,7 +292,7 @@ describe('EntityLineageEditModal', () => {
                 parentEntityDataTypes={[DataClassDataType, SampleTypeDataType]}
                 api={getTestAPIWrapper(jest.fn, {
                     samples: getSamplesTestAPIWrapper(jest.fn, {
-                        getSampleSelectionLineageData: () => Promise.resolve(LINEAGE_DATA_WITH_ALIQUOTS),
+                        getSelectionLineageData: () => Promise.resolve(LINEAGE_DATA_WITH_ALIQUOTS),
                         getSampleOperationConfirmationData: () =>
                             Promise.resolve(new OperationConfirmationData(NONE_ALLOWED_CONFIRMATION_DATA)),
                     }),
@@ -319,7 +319,7 @@ describe('EntityLineageEditModal', () => {
                 parentEntityDataTypes={[DataClassDataType, SampleTypeDataType]}
                 api={getTestAPIWrapper(jest.fn, {
                     samples: getSamplesTestAPIWrapper(jest.fn, {
-                        getSampleSelectionLineageData: () => Promise.resolve(LINEAGE_DATA_WITHOUT_ALIQUOTS),
+                        getSelectionLineageData: () => Promise.resolve(LINEAGE_DATA_WITHOUT_ALIQUOTS),
                         getSampleOperationConfirmationData: () =>
                             Promise.resolve(new OperationConfirmationData(SOME_ALLOWED_CONFIRMATION_DATA)),
                     }),
