+++ packages/components/src/internal/components/samples/actions.ts
@@ -16,7 +16,13 @@
 import { List, Map, OrderedMap } from 'immutable';
 import { ActionURL, Ajax, Domain, Filter, Query, Utils } from '@labkey/api';
 
-import { EntityChoice, EntityDataType, IEntityTypeDetails, IEntityTypeOption } from '../entities/models';
+import {
+    EntityChoice,
+    EntityDataType,
+    EntityParentType,
+    IEntityTypeDetails,
+    IEntityTypeOption
+} from '../entities/models';
 import { deleteEntityType, getEntityTypeOptions } from '../entities/actions';
 import {
     AssayStateModel,
@@ -453,6 +459,48 @@ export const getParentTypeDataForLineage = async (
     }
     return { parentTypeOptions, parentIdData };
 };
+
+export function getLineageEditorUpdateColumns(
+    displayQueryModel: QueryModel,
+    originalParents: Record<string, List<EntityChoice>>
+): { queryInfoColumns: OrderedMap<string, QueryColumn>; updateColumns: List<QueryColumn> } {
+    // model columns should include RowId, Name, and one column for each distinct existing parent (source and/or
+    // sample type) of the selected samples.
+    let queryInfoColumns = OrderedMap<string, QueryColumn>();
+    let updateColumns = List<QueryColumn>();
+    displayQueryModel.queryInfo.columns.forEach((column, key) => {
+        if (key === 'rowid') {
+            queryInfoColumns = queryInfoColumns.set(key, column);
+        } else if (key === 'name') {
+            queryInfoColumns = queryInfoColumns.set(key, column);
+            updateColumns = updateColumns.push(column);
+        }
+    });
+    const parentColumns = {};
+    let parentColIndex = 0;
+    Object.values(originalParents).forEach(sampleParents => {
+        sampleParents.forEach(sampleParent => {
+            const { schema, query } = sampleParent.type;
+            const parentCol = EntityParentType.create({ index: parentColIndex, schema, query }).generateColumn(
+                sampleParent.type.entityDataType.uniqueFieldKey,
+                displayQueryModel.schemaName
+            );
+
+            if (!parentColumns[parentCol.fieldKey]) {
+                parentColumns[parentCol.fieldKey] = parentCol;
+                parentColIndex++;
+            }
+        });
+    });
+    Object.keys(parentColumns)
+        .sort() // Order parent columns so sources come first before sample types, and then alphabetically ordered within the types
+        .forEach(key => {
+            queryInfoColumns = queryInfoColumns.set(key, parentColumns[key]);
+            updateColumns = updateColumns.push(parentColumns[key]);
+        });
+
+    return { queryInfoColumns, updateColumns };
+}
 
 export function getUpdatedLineageRows(
     lineageRows: Array<Record<string, any>>,
